requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
idna==2.6
pytz==2018.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
